+++ src/commands/util/diagnostics-analyzer.ts
